MW-642: Optimization

+++ com.openexchange.userfeedback/src/com/openexchange/userfeedback/internal/FeedbackServiceImpl.java
@@ -249,8 +249,10 @@ public class FeedbackServiceImpl implements FeedbackService {
             writeCon = dbService.getWritableForGlobal(ctxGroup);
             writeCon.setAutoCommit(false);
             List<Long> typeIdsToDelete = getTypeIds(ctxGroup, filter, writeCon);
-            feedBackType.deleteFeedbacks(typeIdsToDelete, writeCon);
-            deleteFeedback(writeCon, feedBackType, filter, ctxGroup);
+            if (typeIdsToDelete.size() > 0) {
+                feedBackType.deleteFeedbacks(typeIdsToDelete, writeCon);
+                deleteFeedback(writeCon, feedBackType, filter, ctxGroup);
+            }
             writeCon.commit();
         } catch (SQLException e) {
             try {
